// angular/src/app/shared/models/company.model.ts
export class Company {
  id?: string;
  name?: string;
  logo_url?: string;
  industry?: string;
  li_updated_at?:string;
  updated_at?: string;
  location?: string;
  location_area?: string;
  employees_linkedin?: string;
  employee_high?: string;
  employee_low?: string;
  twitter_url?: string;
  homepage_url?: string;
  facebook_url?: string;
  crunchbase_url?: string;
  linkedin_url?: string;
  company_type?: string;
  company_location?: string;
  employee_range?:string;
  description?:string;
  date_founded?:string;
  sic_codes?:string;
  street_address?:string;
  city_name?:string;
  state_name?:string;
  zip_code?:string;
  country_name?:string;
  phone?:string;
  ticker?:string;
  annual_cost_of_revenue?:string;
  annual_operating_income?:string;
  revenue?:string;
  fiscal_year?:string;
  annual_gross_profit?:string;
  annual_net_income?:string;
  lastUpdatedDate?:number;
  is_from_tp?:boolean; 
  li_company_id?: string;
  company_name?: string;
  li_slug?: string;
  type?: string;
  website?: string;
  city?: string;
  country?: string;
  region?: string;
  pin_code?: string;
  address1?: string;
  tp_company_id?: string;
  isNew?: boolean;
  linkedin_url_with_name?: string;
  employees?: string;
  employee_on_linkedin?: string;
  specialties?: string;
  followers?: string;
  employee_range_on_linkedin?: any;
  founded?: any;
  crunchbase_url_from_li?: any;
}